@@ -5,12 +5,12 @@ buildscript {
         google()
         jcenter()
         maven { url 'https://oss.sonatype.org/content/repositories/snapshots/' }
+        mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.0'
-
-        classpath 'com.github.dcendents:android-maven-gradle-plugin:2.0'
-        classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1+'
+        classpath 'com.android.tools.build:gradle:3.2.1'
+        classpath 'com.github.dcendents:android-maven-gradle-plugin:2.1'
+        classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.4'
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
@@ -20,6 +20,7 @@ allprojects {
     repositories {
         google()
         jcenter()
+        mavenCentral()
 
         maven {
             url "http://dl.bintray.com/florent37/maven"
@@ -28,12 +29,11 @@ allprojects {
 }
 
 ext {
-    COMPILE_SDK = 27
-    TARGET_SDK = 27
-    minSdkVersion = 14
-    supportVersion = '27.0.2'
-    sourceCompatibilityVersion = JavaVersion.VERSION_1_7
-    targetCompatibilityVersion = JavaVersion.VERSION_1_7
+    COMPILE_SDK=28
+    TARGET_SDK=28
+    MIN_SDK = 14
+    sourceCompatibilityVersion = JavaVersion.VERSION_1_8
+    targetCompatibilityVersion = JavaVersion.VERSION_1_8
 
     libraryVersion = "1.1.1"
 }
