@@ -1,8 +1,9 @@
 package cn.qqtheme.AnimatorSample;
 
 import android.graphics.Path;
-import android.support.annotation.FloatRange;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.FloatRange;
+import androidx.annotation.NonNull;
 
 import java.util.ArrayList;
 import java.util.Random;
